requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai>=1.12.0
faiss-cpu==1.8.0
python-dotenv==1.0.0
numpy>=1.26.0,<2.0
pillow==10.1.0
python-multipart==0.0.20
requests>=2.31.0
ultralytics>=8.0.0
sentence-transformers==3.3.1
